function cleanup and reorg
